--- src/main/java/com/usememo/jugger/global/security/token/domain/KakaoLogoutResponse.java
@@ -1,4 +1,0 @@
-package com.usememo.jugger.global.security.token.domain;
-
-public record KakaoLogoutResponse(long code, String message) {
-}
