--- src/main/java/com/usememo/jugger/global/security/token/domain/LogOutResponse.java
@@ -1,4 +1,0 @@
-package com.usememo.jugger.global.security.token.domain;
-
-public record LogOutResponse(String message) {
-}
