--- src/main/java/com/usememo/jugger/global/security/token/domain/NewTokenResponse.java
@@ -1,4 +1,0 @@
-package com.usememo.jugger.global.security.token.domain;
-
-public record NewTokenResponse(String accessToken) {
-}
