+++ src/main/java/com/usememo/jugger/global/security/token/domain/logOutResponse/KakaoLogoutResponse.java
@@ -1,0 +1,4 @@
+package com.usememo.jugger.global.security.token.domain.logOutResponse;
+
+public record KakaoLogoutResponse(long code, String message) {
+}
