+++ src/main/java/com/usememo/jugger/global/security/token/domain/logOutResponse/LogOutRequest.java
@@ -1,0 +1,4 @@
+package com.usememo.jugger.global.security.token.domain.logOutResponse;
+
+public record LogOutRequest(String refreshToken, String provider) {
+}
