+++ src/main/java/com/usememo/jugger/global/security/token/domain/logOutResponse/LogOutResponse.java
@@ -1,0 +1,4 @@
+package com.usememo.jugger.global.security.token.domain.logOutResponse;
+
+public record LogOutResponse(String message) {
+}
