+++ src/main/java/com/usememo/jugger/global/security/token/domain/oAuthProperties/KakaoOAuthProperties.java
@@ -1,4 +1,4 @@
-package com.usememo.jugger.global.security.token.domain;
+package com.usememo.jugger.global.security.token.domain.oAuthProperties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class KakaoOAuthProperties {
-	private String clientId;
-	private String clientSecret;
-	private String redirectUri;
+    private String clientId;
+    private String clientSecret;
+    private String redirectUri;
 }
