+++ src/main/java/com/usememo/jugger/global/security/token/domain/oAuthProperties/NaverOAuthProperties.java
@@ -1,4 +1,4 @@
-package com.usememo.jugger.global.security.token.domain;
+package com.usememo.jugger.global.security.token.domain.oAuthProperties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
@@ -10,7 +10,6 @@ import lombok.Setter;
 @ConfigurationProperties(prefix = "naver")
 @Getter
 @Setter
-
 public class NaverOAuthProperties {
     private String clientId;
     private String clientSecret;
