+++ src/main/java/com/usememo/jugger/global/security/token/domain/signUp/GoogleSignupRequest.java
@@ -1,23 +1,17 @@
-package com.usememo.jugger.global.security.token.domain;
-
-import com.usememo.jugger.domain.user.entity.User;
+package com.usememo.jugger.global.security.token.domain.signUp;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
-public record KakaoSignUpRequest(
-
+public record GoogleSignupRequest(
         @Schema(description = "사용자 이름", example = "홍길동", required = true)
         String name,
 
         @Schema(description = "이메일 주소", example = "hong@example.com", required = true)
         String email,
 
-        @Schema(description = "로그인 도메인", example = "kakao", required = true)
-        String domain,
-
         @Schema(description = "약관 동의 정보", required = true)
-        Terms terms
+        KakaoSignUpRequest.Terms terms
 
 ) {
     @Schema(name = "Terms", description = "약관 동의 항목들")
