+++ src/main/java/com/usememo/jugger/global/security/token/domain/signUp/NaverSignUpRequest.java
@@ -1,10 +1,9 @@
-package com.usememo.jugger.global.security.token.domain;
+package com.usememo.jugger.global.security.token.domain.signUp;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 public record NaverSignUpRequest(
-
         @Schema(description = "사용자 이름", example = "홍길동", required = true)
         String name,
