+++ src/main/java/com/usememo/jugger/global/security/token/domain/token/NewTokenResponse.java
@@ -1,0 +1,4 @@
+package com.usememo.jugger.global.security.token.domain.token;
+
+public record NewTokenResponse(String accessToken) {
+}
