+++ src/main/java/com/usememo/jugger/global/security/token/domain/token/RefreshToken.java
@@ -1,7 +1,6 @@
-package com.usememo.jugger.global.security.token.domain;
+package com.usememo.jugger.global.security.token.domain.token;
 
 import java.time.Instant;
-import java.util.UUID;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
