+++ src/main/java/com/usememo/jugger/global/security/token/domain/token/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.usememo.jugger.global.security.token.domain;
+package com.usememo.jugger.global.security.token.domain.token;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
