+++ src/main/java/com/usememo/jugger/global/security/token/service/NaverOAuthService.java
@@ -5,13 +5,15 @@ import com.usememo.jugger.domain.user.entity.User;
 import com.usememo.jugger.domain.user.repository.UserRepository;
 import com.usememo.jugger.global.exception.BaseException;
 import com.usememo.jugger.global.exception.ErrorCode;
-import com.usememo.jugger.global.exception.KakaoException;
 import com.usememo.jugger.global.security.JwtTokenProvider;
-import com.usememo.jugger.global.security.token.domain.*;
+import com.usememo.jugger.global.security.token.domain.oAuthProperties.NaverOAuthProperties;
+import com.usememo.jugger.global.security.token.domain.signUp.NaverSignUpRequest;
+import com.usememo.jugger.global.security.token.domain.token.NewTokenResponse;
+import com.usememo.jugger.global.security.token.domain.token.TokenResponse;
+import com.usememo.jugger.global.security.token.domain.userResponse.NaverUserResponse;
 import com.usememo.jugger.global.security.token.repository.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.el.parser.Token;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
