+++ src/main/java/com/usememo/jugger/global/security/token/service/OAuthService.java
@@ -1,6 +1,6 @@
 package com.usememo.jugger.global.security.token.service;
 
-import com.usememo.jugger.global.security.token.domain.NewTokenResponse;
+import com.usememo.jugger.global.security.token.domain.token.NewTokenResponse;
 import org.springframework.http.ResponseEntity;
 import reactor.core.publisher.Mono;
